separated settings component
