+++ src/components/Settings/Settings.module.scss
@@ -1,0 +1,9 @@
+/* Copyright (c) 2024 Massachusetts Institute of Technology */
+/* SPDX-License-Identifier: MIT */
+
+.settings-button {
+  position: absolute;
+  top: 0.5rem;
+  left: 0.5rem;
+  z-index: 1;
+}
